Fixes for Swift 1.2
This commit fixes a number of issues including type changes, ambiguous operators
and some deprecations. Now compiles in XCode 6.4 / Swift 1.2
Also updates to newer versions of the GCDAsyncSocket code to accommodate newer
OS X versions which deprecate some SSL calls

Tested briefly with mosquitto-1.4.2 on the broker side but not guaranteed to be
widely tested or interoperable with other MQTT implementations.